FER_violence_project/requirements.txt
torch
torchvision
opencv-python
tqdm
numpy
Pillow
